+++ handlers/blogs/blogsCommentEditPage.go
@@ -11,7 +11,7 @@ import (
 	"github.com/arran4/goa4web/core"
 	common "github.com/arran4/goa4web/core/common"
 	db "github.com/arran4/goa4web/internal/db"
-	postcountworker "github.com/arran4/goa4web/internal/postcountworker"
+	postcountworker "github.com/arran4/goa4web/workers/postcountworker"
 	"github.com/gorilla/mux"
 )
