+++ handlers/linker/linkerAdminQueuePage.go
@@ -13,7 +13,7 @@ import (
 	common "github.com/arran4/goa4web/core/common"
 	handlers "github.com/arran4/goa4web/handlers"
 	db "github.com/arran4/goa4web/internal/db"
-	searchworker "github.com/arran4/goa4web/internal/searchworker"
+	searchworker "github.com/arran4/goa4web/workers/searchworker"
 
 	"github.com/arran4/goa4web/internal/tasks"
 )
