+++ handlers/linker/linkerCommentsEditPage.go
@@ -11,8 +11,8 @@ import (
 	"github.com/arran4/goa4web/core"
 	common "github.com/arran4/goa4web/core/common"
 	db "github.com/arran4/goa4web/internal/db"
-	postcountworker "github.com/arran4/goa4web/internal/postcountworker"
 	"github.com/arran4/goa4web/internal/tasks"
+	postcountworker "github.com/arran4/goa4web/workers/postcountworker"
 	"github.com/gorilla/mux"
 )
