+++ handlers/linker/linkerShowPage.go
@@ -12,8 +12,8 @@ import (
 	corelanguage "github.com/arran4/goa4web/core/language"
 	handlers "github.com/arran4/goa4web/handlers"
 	db "github.com/arran4/goa4web/internal/db"
-	postcountworker "github.com/arran4/goa4web/internal/postcountworker"
-	searchworker "github.com/arran4/goa4web/internal/searchworker"
+	postcountworker "github.com/arran4/goa4web/workers/postcountworker"
+	searchworker "github.com/arran4/goa4web/workers/searchworker"
 
 	"github.com/arran4/goa4web/config"
