+++ handlers/news/searchResultNewsActionPage.go
@@ -12,7 +12,7 @@ import (
 	"github.com/arran4/goa4web/handlers"
 	handlers "github.com/arran4/goa4web/handlers"
 	db "github.com/arran4/goa4web/internal/db"
-	searchutil "github.com/arran4/goa4web/internal/searchworker"
+	searchutil "github.com/arran4/goa4web/workers/searchworker"
 )
 
 func SearchResultNewsActionPage(w http.ResponseWriter, r *http.Request) {
