+++ handlers/search/searchResultLinkerActionPage.go
@@ -11,7 +11,7 @@ import (
 	handlers "github.com/arran4/goa4web/handlers"
 	hlinker "github.com/arran4/goa4web/handlers/linker"
 	db "github.com/arran4/goa4web/internal/db"
-	searchutil "github.com/arran4/goa4web/internal/searchworker"
+	searchutil "github.com/arran4/goa4web/workers/searchworker"
 
 	"github.com/arran4/goa4web/core"
 	"github.com/arran4/goa4web/internal/tasks"
