+++ handlers/search/searchResultWritingsActionPage.go
@@ -11,7 +11,7 @@ import (
 	handlers "github.com/arran4/goa4web/handlers"
 	hwritings "github.com/arran4/goa4web/handlers/writings"
 	db "github.com/arran4/goa4web/internal/db"
-	searchutil "github.com/arran4/goa4web/internal/searchworker"
+	searchutil "github.com/arran4/goa4web/workers/searchworker"
 
 	"github.com/arran4/goa4web/core"
 	"github.com/arran4/goa4web/internal/tasks"
