+++ handlers/writings/writingsArticleAddPage.go
@@ -10,7 +10,7 @@ import (
 	handlers "github.com/arran4/goa4web/handlers"
 	db "github.com/arran4/goa4web/internal/db"
 	notif "github.com/arran4/goa4web/internal/notifications"
-	searchworker "github.com/arran4/goa4web/internal/searchworker"
+	searchworker "github.com/arran4/goa4web/workers/searchworker"
 	"strings"
 
 	"github.com/arran4/goa4web/core"
