+++ workers/auditworker/worker.go
