+++ workers/logworker/worker.go
