+++ workers/searchworker/insert.go
