+++ workers/searchworker/postupdate.go
