+++ workers/searchworker/worker.go
